consumer/requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.4.4
Click==7.0
dnspython==1.16.0
dominate==2.4.0
eventlet==0.25.1
Flask==1.1.1
Flask-Debug==0.4.3
Flask-SocketIO==4.2.1
greenlet==0.4.15
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.17.4
opencv-contrib-python==4.1.2.30
pathspec==0.6.0
pep8==1.7.1
pycodestyle==2.5.0
python-engineio==3.10.0
python-socketio==4.4.0
regex==2019.11.1
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
u-msgpack-python==2.5.2
visitor==0.1.3
Werkzeug==0.16.0
